from Settings interface was extracted, so it could be used for mocking in tests
